Allow style attribute for img tag
In the old LaTeX CodeCogs editor, the height and width are defined by style
and not seperate height and width attributes.
So style should also be allowed, otherwise it resizing the image won't work.
Also, we now allow more style attributes that could also be set by the LaTeX
editor. So we won't get issue from that again.

This issue was reported by Usman and discussed on Slack