Fix 404 editing drawing question in group question
The $owner_id in Cake is not a real Test UUID somehow for a drawing question.
This causes the issue that Laravel will throw a 404 and this appears
in Bugsnag.
Now we just check for both cases, because it is very time consuming to
fix the Drawing Question itself.

This issue was not blocking, becacuse it is only used for opensource/
not opensource content checking.