Change to blacklist/whitelist, add message+cakepath
Now we can log even more precisely because we can determine the cakepath.
This path is the URL of cake which may trigger this log.
The cakepath may be ommitted and then only the path will be used.

Some initial actions are logged, but in TCP-1381 more actions should
be collected.