fix several issues
o make sure that if a section is deleted, the name is not requested. As it can still be within the school location sections table. Should have been removed, but probably isn't
o make sure that if a subject is searched for, only search it within the sections list of subjects (that is, with the same section_id);