Refactor makeMainElementOfRightType to get rid of switch
The method `makeMainElementOfRightType` in svgShape.js used a switch statement to determine which class to create, which wasn't very DRY and didn't catch errors when someone tried to create an Element of a non-existing type. This has now been solved, as the function uses an Object from constants.js (`elementClassNameForType`) to look up which class to make based on a specified elementType. If the specified type isn't mapped to an existing class it throws an error.