Implement AppAPI version info endpoint
With this endpoint, the apps can check if their current version is allowed
by the server or is about to expire. If it expires, a timestamp is send
which indicates the deadline.
In order to make this possible, the code is refactored to return the Carbon date
and not the iso format string.

Also some potential type confusions / bugs are fixed.