Add versions 0.0.1 and 0.0.2 for testing purposes
These versions have never existed so they won't cause any problems
when we define them.
This allows the app developers to end-to-end test the update mechanism.
Without these versions, we can't test the 'needs update' flow.

These versions can be removed after we have deployed the apps
with the splashscreen.