Add AppApi current date route
This route is used by the apps to query the current date.
The local clock can't be trusted since it can be incorrect and therefore
the wrong tlckey is computed.
Now we allow the app to query the correct current date such that the computation can never be wrong
(taking into account that the current date can shift during execution).